Completed authentication for Genre CRUD operations

+++ RollTheDice/src/main/java/com/skilldistillery/rollthedice/services/GenreServiceImpl.java
@@ -7,13 +7,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.skilldistillery.rollthedice.entities.Genre;
+import com.skilldistillery.rollthedice.entities.User;
 import com.skilldistillery.rollthedice.repositories.GenreRepository;
+import com.skilldistillery.rollthedice.repositories.UserRepository;
 
 @Service
 public class GenreServiceImpl implements GenreService {
 	
 	@Autowired
 	private GenreRepository genreRepo;
+	
+	@Autowired
+	private UserRepository userRepo;
 
 	@Override
 	public List<Genre> index() {
@@ -23,6 +28,7 @@ public class GenreServiceImpl implements GenreService {
 
 	@Override
 	public Genre show(int id) {
+
 		Optional<Genre> op = genreRepo.findById(id);
 		if(op.isPresent()) {
 			return op.get();
@@ -31,8 +37,9 @@ public class GenreServiceImpl implements GenreService {
 	}
 	
 	@Override
-	public Genre create(Genre genre) {
-		if(genre != null) {
+	public Genre create(Genre genre, String username) {
+		User user = userRepo.findByUsername(username);
+		if(genre != null && user.getRole().equals("ROLE_ADMIN")) {
 			return genreRepo.saveAndFlush(genre);
 		}
 		return null;
@@ -42,10 +49,12 @@ public class GenreServiceImpl implements GenreService {
 	
 
 	@Override
-	public Genre update(Genre genre, int gid) {
-		Optional<Genre> opt =genreRepo.findById(gid);
+	public Genre update(Genre genre, int gid, String username) {
+		User user = userRepo.findByUsername(username);
+		
+		Optional<Genre> opt = genreRepo.findById(gid);
 		Genre managed = null;
-		if(opt.isPresent()) {
+		if(opt.isPresent() && user.getRole().equals("ROLE_ADMIN")) {
 			managed = opt.get();
 			managed.setDescription(genre.getDescription());
 			managed.setGames(genre.getGames());
@@ -59,11 +68,13 @@ public class GenreServiceImpl implements GenreService {
 
 
 	@Override
-	public boolean delete(int id) {
+	public boolean delete(int id, String username) {
+		User user = userRepo.findByUsername(username);
+
 		Optional<Genre> op = genreRepo.findById(id);
 		boolean deleted = false;
 		
-		if(op.isPresent()) {
+		if(op.isPresent() && user.getRole().equals("ROLE_ADMIN")) {
 			genreRepo.deleteById(id);
 			deleted = true;
 		}
